First swag at adding leaflet map
Leaflet map function needs to be wired in to es6/jspm module
style. According to Leaflet docs, the map container needs to have a
defined height. We're probably going to need to do some hacky things
with jquery to get this to work with bootstrap, based on what i'm seeing online.