Dynamically adjust size of leaflet map container
Leaflet map container needs a defined map.  Because the site is
responsive, we need to account for the height of the container
changing. Added functionality to dynamically change the height in pixels
of the leaflet map container's parent (the leaflet container's height is
a percentage).  On window resize the parent's height is adjusted by an
event listener.